# toptal/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Restaurant Review App
# during dev, should point to your local machine
host: localhost:3001
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
  /restaurants:
    get:
      tags:
        - restaurants
      x-swagger-router-controller: restaurant
      description: Returns all restaurants
      # used as the method name of the controller
      operationId: getRestaurants
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetRestaurantListResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - restaurants
      description: Creates restaurant
      parameters:
        - in: body
          name: review
          description: The restaurant to create.
          schema:
            $ref: "#/definitions/RestaurantNew"
      responses:
        "201":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetRestaurantListResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  RestaurantNew:
    type: object
    properties:
      userId:
        type: integer
        description: userId for owner
      name:
        type: string
  RestaurantGet:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  GetRestaurantListResponse:
    required:
      - restaurants
    properties:
      # The array of movies
      restaurants:
        type: array
        items:
          $ref: "#/definitions/RestaurantGet"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string